requirements.txt
numpy
pandas
pdfminer.six
Flask
seaborn
matplotlib
plotly
scikit-learn
PyPDF2
